+++ tests/favicon.test.ts
@@ -1,4 +1,5 @@
 import { assertSiteSnapshot, build, getSite } from "./utils.ts";
+import svgo from "../plugins/svgo.ts";
 import favicon from "../plugins/favicon.ts";
 
 Deno.test("favicon plugin", async (t) => {
@@ -6,6 +7,7 @@ Deno.test("favicon plugin", async (t) => {
     src: "favicon",
   });
 
+  site.use(svgo());
   site.use(favicon({
     cache: false,
   }));