+++ CHANGELOG.md
@@ -21,7 +21,7 @@ Any BREAKING CHANGE between minor versions will be documented here in upper case
   - Added some missing UI configurations
 - `imagick` plugin: bug removing duplicated entries.
 - `mdx` plugin: use `remark-gfm@3.0.1` while mdx-js is not upgraded.
-- Updated dependencies: `std`, `deno-dom`, `esbuild`, `imagick`, `markdown-it`, `preact`.
+- Updated dependencies: `std`, `deno-dom`, `esbuild`, `imagick`, `markdown-it`, `preact`, `postcss`.
 
 ## [1.19.0] - 2023-09-25
 ### Added